add to savings goal

+++ app/src/main/java/com/vesna/roundup/presentation/roundup/RoundUpState.kt
@@ -1,3 +1,3 @@
 package com.vesna.roundup.presentation.roundup
 
-class RoundUpState(val roundUp: Double?) // TODO add loading
+class RoundUpState(val roundUp: Int?) // TODO add loading
